+++ Changelog.md
@@ -2,7 +2,8 @@ Csuper
 ======
 
 4.0.0
-* Add a function to transform integer to yes or no
+* Change the dispplay of a game configuration with yes or no instead of 1 and 0
+* Correction of a bug which display the distrubutor when there is no distributor
 
 2.4.0
 
@@ -14,6 +15,7 @@ Libcsuper
 
 4.0.0
 * Add a function to transform integer to yes or no
+* Add a function which calculate the index of the array from the position in the game
 
 2.4.0